ee/cluster-bootstrap/requirements.txt
attrs==21.2.0
aws-cdk.assets==1.125.0
aws-cdk.aws-applicationautoscaling==1.125.0
aws-cdk.aws-autoscaling==1.125.0
aws-cdk.aws-autoscaling-common==1.125.0
aws-cdk.aws-certificatemanager==1.125.0
aws-cdk.aws-cloudformation==1.125.0
aws-cdk.aws-cloudwatch==1.125.0
aws-cdk.aws-codeguruprofiler==1.125.0
aws-cdk.aws-codestarnotifications==1.125.0
aws-cdk.aws-ec2==1.125.0
aws-cdk.aws-ecr==1.125.0
aws-cdk.aws-ecr-assets==1.125.0
aws-cdk.aws-efs==1.125.0
aws-cdk.aws-eks==1.125.0
aws-cdk.aws-elasticloadbalancing==1.125.0
aws-cdk.aws-elasticloadbalancingv2==1.125.0
aws-cdk.aws-events==1.125.0
aws-cdk.aws-iam==1.125.0
aws-cdk.aws-kms==1.125.0
aws-cdk.aws-lambda==1.125.0
aws-cdk.aws-logs==1.125.0
aws-cdk.aws-opensearchservice==1.125.0
aws-cdk.aws-route53==1.125.0
aws-cdk.aws-s3==1.125.0
aws-cdk.aws-s3-assets==1.125.0
aws-cdk.aws-sam==1.125.0
aws-cdk.aws-secretsmanager==1.125.0
aws-cdk.aws-signer==1.125.0
aws-cdk.aws-sns==1.125.0
aws-cdk.aws-sqs==1.125.0
aws-cdk.aws-ssm==1.125.0
aws-cdk.cloud-assembly-schema==1.125.0
aws-cdk.core==1.125.0
aws-cdk.custom-resources==1.125.0
aws-cdk.cx-api==1.125.0
aws-cdk.lambda-layer-awscli==1.125.0
aws-cdk.lambda-layer-kubectl==1.125.0
aws-cdk.region-info==1.125.0
cattrs==1.8.0
constructs==3.3.156
jsii==1.35.0
publication==0.0.3
python-dateutil==2.8.2
PyYAML==5.4.1
six==1.16.0
typing-extensions==3.10.0.2
